backend/auth/requirements.txt
# Dependencias

fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
jinja2
email-validator
